quepistasis/requirements.txt
Package                Version
---------------------- -------------
absl-py                1.0.0
aiofile                3.7.4
aiohttp                3.8.1
aiosignal              1.2.0
alabaster              0.7.8
appdirs                1.4.4
async-timeout          4.0.2
attrs                  19.3.0
autograd               1.3
Automat                0.8.0
autoray                0.3.2
azure-core             1.24.2
azure-identity         1.10.0
azure-quantum          0.25.218240
azure-storage-blob     12.13.0
Babel                  2.6.0
blinker                1.4
cachetools             5.0.0
caio                   0.9.6
certifi                2019.11.28
cffi                   1.15.0
chardet                3.0.4
charset-normalizer     2.1.0
chex                   0.1.1
click                  8.1.3
cloud-init             20.1
colorama               0.4.3
command-not-found      0.3
configobj              5.0.6
constantly             15.1.0
cryptography           2.8
cvxpy                  1.2.1
cycler                 0.11.0
Cython                 0.29.30
dbus-python            1.2.16
decorator              5.1.1
Deprecated             1.2.13
dill                   0.3.5.1
dimod                  0.11.3
diskcache              5.4.0
distro                 1.4.0
distro-info            0.23ubuntu1
dm-tree                0.1.6
docplex                2.23.222
docutils               0.16
dwave-cloud-client     0.10.1
dwave-drivers          0.4.4
dwave-greedy           0.2.5
dwave-hybrid           0.6.8
dwave-inspector        0.3.0
dwave-inspectorapp     0.3.0
dwave-neal             0.5.9
dwave-networkx         0.8.12
dwave-ocean-sdk        5.3.0
dwave-preprocessing    0.4.0
dwave-system           1.15.0
dwave-tabu             0.4.5
dwavebinarycsp         0.2.0
ecos                   2.0.10
entrypoints            0.3
fasteners              0.17.3
Flask                  2.1.2
flatbuffers            2.0
fonttools              4.31.1
frozenlist             1.3.0
future                 0.18.2
gitdb                  4.0.9
GitPython              3.1.27
homebase               1.0.1
httplib2               0.14.0
hyperlink              19.0.0
idna                   2.8
imagesize              1.2.0
imbalanced-learn       0.9.1
imblearn               0.0
importlib-metadata     4.12.0
incremental            16.10.1
isodate                0.6.1
itsdangerous           2.1.2
jax                    0.3.4
jaxlib                 0.3.2
Jinja2                 3.1.2
joblib                 1.1.0
jsonpatch              1.22
jsonpointer            2.0
jsonschema             3.2.0
jupyter-client         6.1.12
jupyter-core           4.11.1
keyring                18.0.1
kiwisolver             1.4.0
language-selector      0.1
launchpadlib           1.10.13
lazr.restfulclient     0.14.2
lazr.uri               1.0.3
liac-arff              2.5.0
MarkupSafe             2.1.1
matplotlib             3.5.1
microsoft              0.1.2
minio                  7.1.5
minorminer             0.2.9
more-itertools         4.2.0
mpmath                 1.2.1
msal                   1.18.0
msal-extensions        1.0.0
msrest                 0.7.1
mthree                 0.24.0
multidict              6.0.2
netifaces              0.10.4
networkx               2.7.1
ninja                  1.10.2.3
ntlm-auth              1.5.0
numpy                  1.22.2
oauthlib               3.1.0
olefile                0.46
openml                 0.12.2
opt-einsum             3.3.0
optax                  0.1.1
orjson                 3.7.3
osqp                   0.6.2.post5
packaging              21.3
pandas                 1.4.0
pbr                    5.9.0
penaltymodel           1.0.2
PennyLane              0.24.0
PennyLane-Lightning    0.24.0
PennyLane-qiskit       0.24.0
Pillow                 9.0.1
pip                    22.0.4
plucky                 0.4.3
ply                    3.11
portalocker            2.5.1
prince                 0.7.1
protobuf               3.20.1
psutil                 5.9.1
py                     1.11.0
pyarrow                7.0.0
pyasn1                 0.4.2
pyasn1-modules         0.2.1
pycparser              2.21
pydantic               1.9.1
pygad                  2.16.3
Pygments               2.3.1
PyGObject              3.36.0
PyHamcrest             1.9.0
PyJWT                  1.7.1
pymacaroons            0.13.0
PyNaCl                 1.3.0
pyOpenSSL              19.0.0
pyparsing              2.4.7
pyqir-generator        0.4.2a1
pyqubo                 1.2.0
pyrsistent             0.15.5
pyserial               3.4
PySocks                1.7.1
python-apt             2.0.0
python-constraint      1.4.0
python-dateutil        2.8.2
python-debian          0.1.36ubuntu1
pytz                   2022.1
PyYAML                 5.3.1
pyzmq                  19.0.2
qdldl                  0.1.5.post2
qiskit                 0.37.0
qiskit-aer             0.10.4
qiskit-ibmq-provider   0.19.2
qiskit-ignis           0.7.1
qiskit-ionq            0.3.5
qiskit-optimization    0.4.0
qiskit-qir             0.1.0b12
qiskit-terra           0.21.0
qsharp                 0.25.218240
qsharp-chemistry       0.25.218240
qsharp-core            0.25.218240
quask                  1.0.0b0
rectangle-packer       2.0.1
requests               2.28.1
requests-ntlm          1.1.0
requests-oauthlib      1.3.1
requests-unixsocket    0.2.0
retry                  0.9.2
retworkx               0.11.0
roman                  2.0.0
scikit-learn           1.1.1
scipy                  1.8.0
scs                    3.2.0
SecretStorage          2.3.1
semantic-version       2.6.0
service-identity       18.1.0
setuptools             62.6.0
simplejson             3.16.0
six                    1.16.0
smmap                  5.0.0
Sphinx                 1.8.5
sphinx-rtd-theme       1.0.0
ssh-import-id          5.10
stevedore              3.5.0
symengine              0.9.2
sympy                  1.10.1
systemd-python         234
threadpoolctl          3.1.0
toml                   0.10.2
toolz                  0.11.2
tornado                6.2
tqdm                   4.64.0
traitlets              5.3.0
tweedledum             1.1.1
Twisted                18.9.0
typing_extensions      4.1.1
ubuntu-advantage-tools 20.3
ufw                    0.36
unattended-upgrades    0.1
urllib3                1.25.8
wadllib                1.3.3
websocket-client       1.3.3
websockets             10.3
Werkzeug               2.1.2
wheel                  0.37.1
wrapt                  1.14.1
xmltodict              0.12.0
yarl                   1.7.2
zipp                   1.0.0
zope.interface         4.7.1
